pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot==20.7",
    "schedule>=1.2.2",
    "streamlit>=1.46.1",
]
